Move load_diagnostic_data.py to load_run_data.py
