+++ workflows/prognostic_run_diags/fv3net/diagnostics/prognostic_run/__init__.py
@@ -1,5 +1,5 @@
 from .config import get_verification_entries
-from .load_diagnostic_data import load_dycore, load_physics
+from .load_run_data import load_dycore, load_physics
 
 
 __all__ = dir()
